Split run test to avoid timeouts
