+++ tests/run-with-compiler/i3876-b.scala
@@ -4,8 +4,11 @@ object Test {
   def main(args: Array[String]): Unit = {
     val x: Expr[Int] = '(3)
 
-    val f: Expr[Int => Int] = '{ (x: Int) => x + x }
-    println(f(x).run)
-    println(f(x).show)
+    val f2: Expr[Int => Int] = '{
+      def f(x: Int): Int = x + x
+      f
+    }
+    println(f2(x).run)
+    println(f2(x).show)
   }
 }
