+++ tests/run-with-compiler/i3876-d.check
@@ -1,0 +1,8 @@
+6
+{
+  val x$1: Int = 3
+  /* inlined from Test*/ 
+    {
+      x$1.+(x$1)
+    }
+}
