+++ tests/run-with-compiler/i3876.check
@@ -5,35 +5,3 @@
     x$1.+(x$1)
   }
 }
-6
-{
-  val x$1: Int = 3
-  {
-    def f(x: Int): Int = x.+(x)
-    f(x$1)
-  }
-}
-6
-{
-  val x$1: Int = 3
-  {
-    val f: 
-      Function1[Int, Int] 
-        {
-          def apply(x: Int): Int
-        }
-     = 
-      {
-        (x: Int) => x.+(x)
-      }
-    (f: (x: Int) => Int).apply(x$1)
-  }
-}
-6
-{
-  val x$1: Int = 3
-  /* inlined from Test*/ 
-    {
-      x$1.+(x$1)
-    }
-}
