+++ tests/run-with-compiler/quote-run-b.check
@@ -1,0 +1,3 @@
+
+lambda(4)
+lambda(5)
