+++ tests/run-with-compiler/quote-run-b.scala
@@ -1,0 +1,17 @@
+
+import dotty.tools.dotc.quoted.Runners._
+
+import scala.quoted._
+
+object Test {
+  def main(args: Array[String]): Unit = {
+    val lambdaExpr = '{
+      (x: Int) => println("lambda(" + x + ")")
+    }
+    println()
+
+    val lambda = lambdaExpr.run
+    lambda(4)
+    lambda(5)
+  }
+}
