+++ tests/run-with-compiler/quote-run-c.check
@@ -1,0 +1,4 @@
+Foo
+false
+Bar
+class Quoted$A$1
